refactor: refactor layout
