+++ apps/web/app/page.tsx
@@ -1,11 +1,11 @@
-import HeroSection from "@/components/Feature/landing/Hero";
+import HeroSection from "@/components/landing/Hero";
 import TestimonialSection from "@/components/landing/Testimonial";
 import NavBar from "@/components/Layout/Navbar";
 import { PageContainer, Main } from "./styles";
-import ValueStatement from "@/components/Feature/landing/ValueStatement";
+import ValueStatement from "@/components/landing/ValueStatement";
 import Footer from "@/components/Layout/Footer";
-import WatchingSteps from "@/components/Feature/landing/WatchingSteps";
-import DiscoverContent from "@/components/Feature/landing/DiscoverContent";
+import WatchingSteps from "@/components/landing/WatchingSteps";
+import DiscoverContent from "@/components/landing/DiscoverContent";
 import CallToAction from "@/components/landing/CallToAction";
 
 export default function Home() {
@@ -16,11 +16,10 @@ export default function Home() {
         <HeroSection />
         <TestimonialSection />
         <DiscoverContent />
+        <WatchingSteps />
+        <ValueStatement />
+        <CallToAction />
       </Main>
-
-      <WatchingSteps />
-      <ValueStatement />
-      <CallToAction />
       <Footer />
     </PageContainer>
   );
