+++ apps/web/components/landing/DiscoverContent/index.tsx
