+++ apps/web/components/landing/WatchingSteps/index.tsx
@@ -20,7 +20,7 @@ import {
 } from "./styles";
 import Image from "next/image";
 import { useTranslation } from "react-i18next";
-import Steps from "../../../../assets/images/steps.png";
+import Steps from "../../../assets/images/steps.png";
 
 export default function WatchingSteps() {
   const { t } = useTranslation();
