+++ apps/web/components/landing/WatchingSteps/styles.ts
